feat(auth): derive serde for HeimaMultiSignature
Move serde serialization and deserialization for HeimaMultiSignature
directly into the enum using serde attributes. Remove the
HeimaMultiSignatureSerde wrapper and update all usages and tests to use
the enum directly. This simplifies the codebase and ensures consistent
serialization for multi-signature types.
