+++ tee-worker/omni-executor/executor-primitives/Cargo.toml
@@ -16,7 +16,7 @@ tracing = { workspace = true }
 
 executor-crypto = { workspace = true }
 heima-primitives = { workspace = true }
-sp-core = { workspace = true }
+sp-core = { workspace = true, features = ["serde"] }
 
 [dev-dependencies]
 serde_json = { workspace = true }
